Refactor: Decorators - Move decorators to their own files
- Move 'BaseDecorator', 'CapitalizeDecorator', and 'TrimmerDecorator'
in 'nameable.rb' to their own files.
