+++ lib/class_decorators/nameable.rb
@@ -3,28 +3,3 @@ class Nameable
     raise NotImplementedError, "#{self.class} has not implemented method '#{__method__}'"
   end
 end
-
-class BaseDecorator < Nameable
-  attr_accessor :nameable
-
-  def initialize(nameable)
-    super()
-    @nameable = nameable
-  end
-
-  def correct_name
-    @nameable.correct_name
-  end
-end
-
-class CapitalizeDecorator < BaseDecorator
-  def correct_name
-    @nameable.correct_name.capitalize
-  end
-end
-
-class TrimmerDecorator < BaseDecorator
-  def correct_name
-    @nameable.correct_name[0...10] unless @nameable.correct_name.length <= 10
-  end
-end
