+++ lib/class_decorators/trimmer_decorator.rb
@@ -1,0 +1,7 @@
+require_relative 'base_decorator'
+
+class TrimmerDecorator < BaseDecorator
+  def correct_name
+    @nameable.correct_name[0...10] unless @nameable.correct_name.length <= 10
+  end
+end
